Legger til request objekter
