+++ src/main/kotlin/no/nav/helsearbeidsgiver/dialogporten/domene/Dialog.kt
@@ -1,9 +1,16 @@
+@file:UseSerializers(UuidSerializer::class)
+
 package no.nav.helsearbeidsgiver.dialogporten.domene
 
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.UseSerializers
+import no.nav.helsearbeidsgiver.utils.json.serializer.UuidSerializer
+import no.nav.helsearbeidsgiver.utils.wrapper.Orgnr
+import java.util.UUID
 
 @Serializable
 data class Dialog(
+    val id: UUID? = null,
     val serviceResource: String,
     val party: String,
     val externalReference: String,
@@ -11,4 +18,14 @@ data class Dialog(
     val content: Content,
     val transmissions: List<Transmission>,
     val isApiOnly: Boolean? = true,
+)
+
+@Serializable
+data class CreateDialogRequest(
+    val orgnr: Orgnr,
+    val title: String,
+    val summery: String,
+    val externalReference: String,
+    val isApiOnly: Boolean = true,
+    val transmissions: List<Transmission>,
 )
