+++ src/test/kotlin/no/nav/helsearbeidsgiver/dialogporten/DialogportenClientTest.kt
@@ -6,6 +6,8 @@ import io.kotest.matchers.shouldBe
 import io.ktor.http.HttpStatusCode
 import no.nav.helsearbeidsgiver.dialogporten.domene.ApiAction
 import no.nav.helsearbeidsgiver.dialogporten.domene.Content
+import no.nav.helsearbeidsgiver.dialogporten.domene.CreateApiActionRequest
+import no.nav.helsearbeidsgiver.dialogporten.domene.CreateTransmissionRequest
 import no.nav.helsearbeidsgiver.dialogporten.domene.Transmission
 import no.nav.helsearbeidsgiver.dialogporten.domene.create
 import java.util.UUID
@@ -24,8 +26,7 @@ class DialogportenClientTest :
             val dialogId = UUID.randomUUID()
             dialogportenClient.addTransmission(
                 dialogId,
-                Transmission(
-                    id = UUID.randomUUID(),
+                CreateTransmissionRequest(
                     type = Transmission.TransmissionType.Information,
                     extendedType = "extendedType",
                     sender = Transmission.Sender(actorType = "actorType"),
@@ -52,8 +53,8 @@ class DialogportenClientTest :
             val dialogportenClient = mockDialogportenClient(HttpStatusCode.NoContent)
             val dialogId = UUID.randomUUID()
             val apiActions =
-                ApiAction(
-                    action = ApiAction.Action.READ.value,
+                CreateApiActionRequest(
+                    action = ApiAction.Action.READ,
                     name = "name",
                     endpoints = emptyList(),
                 )
